% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics.R
\name{agg_score}
\alias{agg_score}
\title{Aggregate Scores}
\usage{
agg_score(scores.df)
}
\arguments{
\item{Data}{frame containing multiple scoring procedures.}
}
\value{
Summarize Reference Gradient, Reference Categorical, and 
All Gradient scoring procedures.
}
\description{
Aggregate Scores
}
