% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics.R
\name{all_gradient}
\alias{all_gradient}
\title{Scoring Method: All Site Classes Gradient}
\usage{
all_gradient(metrics.df, index.metrics)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station}
}
\value{
Scores the samples.
}
\description{
Scoring Method: All Site Classes Gradient
}
