% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics _final.R
\name{all_gradient2}
\alias{all_gradient2}
\title{Scoring Method: All Gradient2}
\usage{
all_gradient2(metrics.df)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station}
}
\value{
Scores the samples.
}
\description{
Scoring Method: All Gradient2
}
