% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics2.R
\name{bde}
\alias{bde}
\title{Balanced Discrimination Efficiency}
\usage{
bde(scored.df, bioregion)
}
\arguments{
\item{scored.df}{= final scored index.}

\item{bioregion}{= the bioregion of interest.}
}
\value{
Update
}
\description{
Balanced Discrimination Efficiency
}
