% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics2.R
\name{best.distribution.H}
\alias{best.distribution.H}
\title{Best Distribution H}
\usage{
best.distribution.H(metrics.df, metric.summary, scores.df, m.c,
  range_var = TRUE, redund = FALSE, best.metrics)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station.}
}
\value{
Scores the samples.
}
\description{
Best Distribution H
}
