% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics2.R
\name{best.distribution}
\alias{best.distribution}
\title{Best Distribution}
\usage{
best.distribution(metrics.df, metric.summary, scores.df, m.c,
  range_var = TRUE)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station.}
}
\value{
Scores the samples.
}
\description{
Best Distribution
}
