% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics2.R
\name{best.distribution3}
\alias{best.distribution3}
\title{Best Distribution 3}
\usage{
best.distribution3(metrics.df, metric.summary, scores.df, m.c, Family = FALSE,
  master, range_var = TRUE)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station.}
}
\value{
Scores the samples.
}
\description{
Best Distribution 3
}
