% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics.R
\name{best_metrics}
\alias{best_metrics}
\title{Select the best metrics available}
\usage{
best_metrics(Minus_Redund.df, best_type = 55, more_metrics = 60)
}
\arguments{
\item{Minus_Redund.df}{= data frame after redundancy analysis}

\item{best_type}{=}

\item{more_metrics}{=}
}
\value{
Determines the best metric to represent each metric type
(i.e. Diversity, Tolerance, Composition, FFG, Habit) and includes any additional
metrics if the metrics discrimination efficiency value is >= 75.
}
\description{
Select the best metrics available
}
