% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics.R
\name{build_ibi}
\alias{build_ibi}
\title{Build a new IBI}
\usage{
build_ibi(Info, Long, Level = "FAMILY", Category,
  all.metrics = all_metrics_div_rare)
}
\arguments{
\item{Info}{= taxonomic information}

\item{Long}{= taxonomic data in a long data format}

\item{Level}{= the lowest taxonomic level that metrics should be calculated}

\item{Category}{= each sampling event assigned a site category (i.e. Reference,
Mixed, or Degraded)}

\item{all.metrics}{= the function that should be used to calculate
all of the metrics}
}
\value{
Build an IBI from scratch.
}
\description{
Build a new IBI
}
