% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_metrics.R
\name{calc_metric}
\alias{calc_metric}
\title{Should the Metric be Calculated?}
\usage{
calc_metric(metric.name, metrics.vec)
}
\arguments{
\item{metric.name}{= the metric in question.}

\item{metrics.vec}{= the vector of specified metrics to calculate.}
}
\value{
This function is used within the all_metrics function and is not 
necissarily useful outside of those functions. The purpose of the function is
to check that the metric in question (metric.name) is present within the
specifiec list of metrics to calculate (metrics.vec). If metrics.vec is 
set to "ALL" then all of the metrics will be calculated. This function was
created to speed up the calculation of metrics. Often calculating all of the
metrics is only useful during the exploratory stage of developing an index. 
Calculating all of the metrics, especially for large data sets, can be time 
consuming. This function helps to skip over metrics that the user is not 
interested in.
}
\description{
Should the Metric be Calculated?
}
