% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sensitivity.R
\name{chunk_sensitivity}
\alias{chunk_sensitivity}
\title{Chunk Sensitivity}
\usage{
chunk_sensitivity(metrics.df, upper.class = "REF", lower.class = "SEV",
  method)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station with site
a column of site classes defined by environmental variables.}

\item{upper.class}{= the site class that represents the better condition.}

\item{lower.class}{= the site class that represents the poorer condition.}

\item{method}{= the sensitivity function to be used during the assessment.}
}
\value{
Determines the threshold at which a metric best categorizes
two defined environmental conditions.
}
\description{
Chunk Sensitivity
}
