% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_metrics.R
\name{corr_taxa}
\alias{corr_taxa}
\title{Taxa Correlations}
\usage{
corr_taxa(Long, Level, value = 0.5)
}
\arguments{
\item{Long}{= Long data formate containing taxonomic counts.}

\item{Level}{= The taxonomic level used in the analysis.}

\item{Coef}{= The correlation coeficient to be used. The value will be used
as the positive and negative threshold.}
}
\value{
Taxa positively or negatively correlated with each other.  These
taxa may work well togther as a percent metric.
}
\description{
Taxa Correlations
}
