% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrimination_efficiency.R
\name{d_e}
\alias{d_e}
\title{Discrimination Efficiency}
\usage{
d_e(deg.df, quant.df)
}
\arguments{
\item{deg.df}{= a data frame of only the lower.class values.}

\item{quant.df}{= Data frame containing upper.class quantile values.}
}
\value{
Determines the threshold at which a metric best categorizes
reference and degraded stations.
}
\description{
Discrimination Efficiency
}
