% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_parametric_test.R
\name{dunn_table}
\alias{dunn_table}
\title{Post hoc Dunn's test p-value and test Statistic}
\usage{
dunn_table(Prep_Taxa, Info, Tab_Project, Tab_Event)
}
\arguments{
\item{Prep_Taxa}{= Taxanomic counts in a long data format.}

\item{Info}{= Taxanomic attributes table.}

\item{Tab_Project}{= Table containing Project_ID and agency name.}

\item{Tab_Event}{= Table containing Project_ID and Event_ID}
}
\value{
Creates a table of the post hoc Dunn's testoutput.
The table includes the EVENT_ID, test statistic, and the p-values
comparing agencies.
}
\description{
Post hoc Dunn's test p-value and test Statistic
}
