% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{fill_taxa}
\alias{fill_taxa}
\title{Fill empty rows with lowest level of}
\usage{
fill_taxa(taxa.long)
}
\arguments{
\item{taxa.long}{= A long data frame containing multiple levels of the 
taxonomic hierarchy.}
}
\value{
Fills in blank spaces with the previous lowest level of taxonomic
identification.
}
\description{
Fill empty rows with lowest level of
}
