% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics2.R
\name{final.score}
\alias{final.score}
\title{Final Score}
\usage{
final.score(final.metrics.summary, metrics.df, scores.df, m.c)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station.}
}
\value{
Scores the samples.
}
\description{
Final Score
}
