% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics.R
\name{ibi_rating}
\alias{ibi_rating}
\title{BIBI Rating}
\usage{
ibi_rating(Scored.df, index = 2011)
}
\arguments{
\item{index}{= 2011 or 2016 rating procedure.}

\item{metrics.df}{= data frame with each metric scored based on}
}
\value{
The index rating for each site. The rating is based on
the average score of the metrics in the index.
}
\description{
BIBI Rating
}
