% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_jackknife_functions.R
\name{jack_sim}
\alias{jack_sim}
\title{Delete-d Jackknife Calculations}
\usage{
jack_sim(train.samp, ref_deg, bioregion, zero.null, Fam, Master, redund.df,
  m.c = metric.class, metric_types = FALSE, method)
}
\arguments{
\item{bioregion}{= the bioregion of interest}

\item{zero.null}{= If TRUE, all zeros are converted to NA for all metrics
with > 5% of Reference values equal to zero. If FALSE, metric values are
unaltered.}

\item{Fam}{= If TRUE, all family metrics are grouped and scored together 
prior to the final scoring procedure.  If FALSE, family metrics are grouped
with composition metrics.}

\item{Master}{= a data frame containing taxonomic traits and attributes.}

\item{redund.df}{= IF TRUE, metrics are evaluated for redundancy. If FALSE,
a redundancy analysis is not preformed.}

\item{m.c}{= a data frame containing all of the metrics and thier associated
metric category.}

\item{method}{= specify the method being preformed. FIGURE OUT A BETTER WAY!!!!}

\item{metrics.df}{= a data frame containing the raw metric values. Only the
metrics selected for the final index are included.}

\item{runs}{= number of jackknife resampling events.}

\item{keep}{= the proportion of the reference population that should be kept.}
}
\value{
Recomputes the final IBI threshold after removing a specified
percentage of the data.
}
\description{
Delete-d Jackknife Calculations
}
