% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_jackknife_functions.R
\name{jack_summarize}
\alias{jack_summarize}
\title{Delete-d Jackknife Select Training Data Set}
\usage{
jack_summarize(jack.sim, bioregion, ref.df, deg.df)
}
\arguments{
\item{jack.sim}{= output from the jack_sim function.}

\item{bioregion}{= the bioregion or region of interest.}

\item{ref.df}{= a data frame containing only Reference samples.}

\item{deg.df}{= a data frame containing only Degraded samples.}
}
\value{
Summarizes the Validation-set CE, Training-set CE, and BSP.
}
\description{
Delete-d Jackknife Select Training Data Set
}
