% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_jackknife_functions.R
\name{jack_train_samp}
\alias{jack_train_samp}
\title{Delete-d Jackknife Select Training Data Set}
\usage{
jack_train_samp(ref.df, deg.df, ref_deg, runs, keep)
}
\arguments{
\item{ref.df}{= a data frame containing only Reference samples.}

\item{deg.df}{= a data frame containing only Degraded samples.}

\item{ref_deg}{= a data frame containing only Reference and Degraded samples.}

\item{runs}{= number of jackknife resampling events.}

\item{keep}{= the proportion of the reference population that should be kept.}
}
\value{
Performes a Delete-d Jackknife to create a unique subset of the orginal
data set for the number of specified runs.
}
\description{
Delete-d Jackknife Select Training Data Set
}
