% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_parametric_test.R
\name{kd_table}
\alias{kd_table}
\title{Kruskal Wallis p-value and Dunn's test}
\usage{
kd_table(Prep_Taxa, Info, Tab_Project, Tab_Event)
}
\arguments{
\item{Prep_Taxa}{= Taxanomic counts in a long data format.}

\item{Info}{= Taxanomic attributes table.}

\item{Tab_Project}{= Table containing Project_ID and agency name.}

\item{Tab_Event}{= Table containing Project_ID and Event_ID}
}
\value{
Creates a table of the Kruskal Wallis and Dunn's test output.
The table includes the EVENT_ID, test statistic, the
Kruskal Wallis p-values, and the Dunn's test p-values comparing agencies.
}
\description{
Kruskal Wallis p-value and Dunn's test
}
