% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics2.R
\name{metric_class_break}
\alias{metric_class_break}
\title{Metric Class Break}
\usage{
metric_class_break(metric_class, class)
}
\arguments{
\item{class}{= metric class.}

\item{metrics.df}{= data frame of metric values for each station.}
}
\value{
Scores the samples.
}
\description{
Metric Class Break
}
