% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics2.R
\name{metric_class_redund}
\alias{metric_class_redund}
\title{Redundancy analysis by metric class}
\usage{
metric_class_redund(metrics.df, metric.summary, m.c, method = ibi.method)
}
\arguments{
\item{metrics.df}{= a data frame with raw metric scores}

\item{metric.summary}{= metric summary data frame}

\item{m.c}{= data frame containing metric classes}

\item{method}{= the ibi.method specified}
}
\value{
Update
}
\description{
Redundancy analysis by metric class
}
