% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sensitivity.R, R/scoring_metrics2.R
\name{metrics_summary}
\alias{metrics_summary}
\alias{metrics_summary}
\title{Summary of Metric Tests}
\usage{
metrics_summary(metrics.df, bioregion, zero = TRUE)

metrics_summary(metrics.df, bioregion, zero = TRUE)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station with site
a column of site classes defined by environmental variables.}

\item{bioregion}{= the bioregion to perform the analysis.}

\item{metrics.df}{= data frame of metric values for each station.}

\item{bioregion}{= the bioregion of interest.}
}
\value{
Summarizes multiple metric tests into a single table.

Scores the samples.
}
\description{
Summary of Metric Tests

Metrics Summary
}
