% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics2.R
\name{metrics_summary2}
\alias{metrics_summary2}
\title{Metrics Summary}
\usage{
metrics_summary2(metrics.df, bioregion)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station.}

\item{bioregion}{= the bioregion of interest.}
}
\value{
Scores the samples.
}
\description{
Metrics Summary
}
