% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundancy.R
\name{old_redundancy}
\alias{old_redundancy}
\title{Redundancy Analysis}
\usage{
old_redundancy(metrics.df, analysis = "bibi", sensitivity.df,
  method.corr = "spearman", upper.coef = 0.7, lower.coef = -0.7,
  upper.class = NULL, lower.class = NULL)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station}

\item{analysis}{= the method for descriminating between two redundant methods.
Two choices are available. "wilcox" will run a Wilcoxin Rank Sum Test on the
designated upper.class and lower.class for each metric. If two metrics are
redundant the metric with the smaller p-value is kept for further analysis.
"bibi" requires a data from the "ODE" metric sensitivity analysis. The
Descrimination Efficiency threshold from the "ODE" analysis is used to
 discriminate between two redundant metrics.  The metric with the larger
 threshold is kept for further analysis.}

\item{sensitivity.df}{= data frame of metrics scored using the Barbour et al. 1996 method
for testing metric sensitivity}

\item{method.corr}{= correlation method to be used.}

\item{upper.coef}{= the upper correlation coeficient to be used in the
correlation analysis.}

\item{lower.coef}{= the lower correlation coeficient to be used in the
correlation analysis.}

\item{upper.class}{= Only enter this class if using analysis = "wilcox".
The reference class to be used in the the Wilcoxin Rank Sum Test.}

\item{lower.class}{= Only enter this class if using analysis = "wilcox".
The degraded class to be used in the the Wilcoxin Rank Sum Test.}
}
\value{
A data frame of metrics scored using the Barbour et al. 1996 method
for testing metric sensitivity excluding the "weaker" redundant metric(s)
using the interquartiles.
}
\description{
Redundancy Analysis
}
