% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics2.R
\name{old_scoring3}
\alias{old_scoring3}
\title{Old Scoring 3}
\usage{
old_scoring3(metrics.df, bioregion, scoring.method, zero_null = TRUE,
  bound.lim = TRUE, metric.summary = NULL)
}
\value{
Scores the samples.
}
\description{
Old Scoring 3
}
