% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics _final.R
\name{old_scoring4}
\alias{old_scoring4}
\title{Old Scoring 4}
\usage{
old_scoring4(metrics.df, bioregion, zero_null = TRUE, bound.lim = TRUE,
  metric.summary = NULL, scoring.method = "UNEVEN_GRADIENT")
}
\value{
Scores the samples.
}
\description{
Old Scoring 4
}
