% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_metrics.R
\name{pct_attribute}
\alias{pct_attribute}
\title{The percent of a group}
\usage{
pct_attribute(Taxa.df, Info, Group, Group_Level, Level = "FAMILY")
}
\arguments{
\item{Taxa.df}{= Wide data frame format of taxonomic counts.}

\item{Info}{= Taxonomic attributes table.}

\item{Group}{= The taxonomic group to be assessed}

\item{Group_Level}{= The specific level or levels of the group to be assessed.}

\item{Level}{= The taxonomic level used during the assessment.}
}
\value{
The percentage of taxa representing a predefined group. Typically,
this function is used to assess functional feeding group and habits.
}
\description{
The percent of a group
}
