% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_metrics.R
\name{pct_dom1_group}
\alias{pct_dom1_group}
\title{The percent of the most dominant group}
\usage{
pct_dom1_group(Long, Info, Group, Level)
}
\arguments{
\item{Long}{= Long data frame format of taxonomic counts.}

\item{Info}{= Taxonomic attributes table.}

\item{Group}{= The taxonomic group to be assessed}

\item{Level}{= The taxonomic level used during the assessment.}
}
\value{
The percentage of taxa representing by the most dominant (abundant)
group. Typically, this function is used to assess the functional feeding
groups and habits.
}
\description{
The percent of the most dominant group
}
