% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{prarefy}
\alias{prarefy}
\title{Most Probable Rarefied Assemblage}
\usage{
prarefy(Long, Sample = 100)
}
\arguments{
\item{Long}{= Taxa count data}

\item{Sample}{= The rarefied sample size}
}
\value{
Selects the assemblage with the highest probability of being selected
during a random rarefaction.
}
\description{
Most Probable Rarefied Assemblage
}
