% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{prep_rare}
\alias{prep_rare}
\title{Rarify Taxonomic Counts}
\usage{
prep_rare(Long, master, Level = "FAMILY", Sample = 100, pct_un = NULL,
  bibi.standard = FALSE)
}
\arguments{
\item{Long}{= Taxa count data}

\item{Level}{= Taxonomic level to preform rarefaction on.}

\item{Sample}{= The rarefaction subsample size.}

\item{pct_un}{= A threshold can be established to exclude samples that do
have too many unidentified taxa in a sample. Enter the percentage of the
sample that you are comfortable with being unidentified.  If not specified
the function will not excluded any samples.}

\item{bibi.standard}{= Indicate TRUE if you would like to standardize your
taxa following the practices used in the Chessie BIBI.  The default, FALSE,
will not attempt to standardize the taxa.}
}
\value{
Rarified counts
}
\description{
Rarify Taxonomic Counts
}
