% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefaction.R
\name{prep_rrarefy}
\alias{prep_rrarefy}
\title{rrarefied counts in a long data frame}
\usage{
prep_rrarefy(Long.df, Level)
}
\arguments{
\item{Level}{= The taxonomic level to preform the analysis.}

\item{Long}{= Taxa count data}
}
\value{
The vegan package function rrarefy is used to subsample the
orginal assemblage. Level is used to specify the lowest taxonomic resolution.
Rarefaction is performed at the rank specified by Level. The new wide data
 frame is merged with the orginal Long data frame to represent that
 rarified counts and all of the applicable taxonomic ranks.
}
\description{
rrarefied counts in a long data frame
}
