% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics.R
\name{prep_score}
\alias{prep_score}
\title{Assigns Index Rating}
\usage{
prep_score(Scores, Metrics)
}
\arguments{
\item{Scores}{= Scores for each metric in the index.}

\item{Metrics}{= Raw metric scores.  Includes only metrics within the index.}
}
\value{
Prepares a the raw metric values, the metric score, and the sampling
event's rating (i.e. "Very Poor", "Poor", "Fair", "Good", or "Excellent).
}
\description{
Assigns Index Rating
}
