% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sensitivity.R
\name{range_variability}
\alias{range_variability}
\title{Range and Variability Test}
\usage{
range_variability(metrics.df)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station with site
a column of site classes defined by environmental variables.}
}
\value{
Tests that the range of the reference condition is not too low and
that variability is not too high.
}
\description{
Range and Variability Test
}
