% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics.R
\name{rate_samples}
\alias{rate_samples}
\title{Rate Samples}
\usage{
rate_samples(Metrics.df, Final_Metrics.df)
}
\arguments{
\item{Metrics.df}{= data frame with all of the raw metric scores.}

\item{Final_Metrics.df}{= a list of metrics selected/identified as
sensitive to a disturbance gradient.}
}
\value{
The index rating for each sampling event. The metrics are scored on
a scale of 0-100 based on descrimination efficiency values between each of
the designated site classes.  The mean score of all of the metrics for each
sampling event is used to rate sampling station.
}
\description{
Rate Samples
}
