% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_metrics.R
\name{ratio_ept_chiro}
\alias{ratio_ept_chiro}
\title{Ratio of EPT to Chironomidae}
\usage{
ratio_ept_chiro(order.wide, family.wide)
}
\arguments{
\item{order.wide}{= Taxonomic counts aggregated at the order level
classification in a wide data format. Use the wide function to
prepare the data.}

\item{family.wide}{= Taxonomic counts aggregated at the family level
classification in a wide data format. Use the wide function to
prepare the data.}
}
\value{
The ratio of EPT (Ephemeroptera, Plecoptera, and Trichoptera) to
chironomids (Family: Chironomidae). One is added to the count of EPT
and to the count of chironomids to avoid issues associate with the absence
of EPT or chironomids.  If no EPT individuals are observed the ratio
(0/# chironomids) will return 0 despite the number of chironomids observed.
If no chironomids were observed the ratio (# EPT/0) will return an inproper
fraction despite the number of EPT observed. I do not recommend using ratios to
develop indices of biotic integrity.  However, I included this metric becuase it
occurs frequently in the literature.  This metric is not included in the
all_metrics function.
}
\description{
Ratio of EPT to Chironomidae
}
