% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics.R
\name{ref_categorical}
\alias{ref_categorical}
\title{Scoring Method: Reference Categorical}
\usage{
ref_categorical(long.df)
}
\arguments{
\item{long.df}{<- a data frame of samples to be scored.}
}
\value{
Scores the samples.
}
\description{
Scoring Method: Reference Categorical
}
