% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics.R
\name{ref_gradient}
\alias{ref_gradient}
\title{Scoring Method: Reference Gradient}
\usage{
ref_gradient(long.df)
}
\arguments{
\item{long.df}{<- a data frame of samples to be scored.}
}
\value{
Scores the samples.
}
\description{
Scoring Method: Reference Gradient
}
