% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics.R
\name{ref_threshold}
\alias{ref_threshold}
\title{Scoring Method: Reference Threshold}
\usage{
ref_threshold(long.df)
}
\arguments{
\item{long.df}{<- a data frame of samples to be scored.}
}
\value{
Scores the samples.
}
\description{
Scoring Method: Reference Threshold
}
