% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_metrics.R
\name{rich_attribute}
\alias{rich_attribute}
\title{The richness of a group}
\usage{
rich_attribute(taxa.wide, master = BIBI::master, attribute.column,
  attribute.interest, rank = "FAMILY")
}
\arguments{
\item{taxa.wide}{= Wide data frame format of taxonomic counts.}

\item{master}{= Taxonomic attributes table.}

\item{attribute.column}{= The name of the column that contains the
attribute of interest.}

\item{attribute.interest}{= The specific attribute of interest}

\item{rank}{= The taxonomic level used during the assessment.}
}
\value{
The richness of taxa representing a predefined group. Typically,
this function is used to assess functional feeding group and habits.
}
\description{
The richness of a group
}
