% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics _final.R
\name{score_metrics}
\alias{score_metrics}
\title{Score Metrics}
\usage{
score_metrics(metrics.df, metrics.summary, bioregion, scoring.method,
  bound.limits = TRUE)
}
\arguments{
\item{metrics.df}{= a data frame containing raw metric scores.}

\item{metrics.summary}{= a data frame of the output from the function 
metric_summary.}

\item{bioregion}{= the bioregion of interest.}

\item{scoring.method}{= the scoring method of interest.}
}
\value{
Update
}
\description{
Score Metrics
}
