% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics.R
\name{score_samples}
\alias{score_samples}
\title{Score Samples}
\usage{
score_samples(metrics.df, index.metrics, method = "all")
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station.}

\item{index.metrics}{= a data frame containing the metrics selected for
the index.}

\item{method}{= Designates the the method(s) used to score the samples.}
}
\value{
Scores the samples.
}
\description{
Score Samples
}
