% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics2.R
\name{scoring2}
\alias{scoring2}
\title{Scoring 2}
\usage{
scoring2(metrics.df, scores, ms, m.c, Family = FALSE, master = master,
  redund, method = ibi.method)
}
\value{
Scores the samples.
}
\description{
Scoring 2
}
