% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics2.R
\name{select_best}
\alias{select_best}
\title{Select the Best Metrics}
\usage{
select_best(metrics.df, metric.summary, m.c, range_var, redund)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station.}
}
\value{
Select the best metrics from 4 of the 5 metric categories
}
\description{
Select the Best Metrics
}
