% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Metric Sensitivity}
\usage{
sensitivity(metrics.df, upper.class, lower.class, method = "ODE")
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station}

\item{upper.class}{= The site classification that represents better
environmental conditions.}

\item{lower.class}{= The site classification that represents the degraded
environmental conditions.}

\item{method}{= the sensitivity function to be used during the assessment.}
}
\value{
Determines the threshold at which a metric best categorizes
reference and degraded stations.
}
\description{
Metric Sensitivity
}
