% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_pct_taxa.R
\name{seq_pct_taxa}
\alias{seq_pct_taxa}
\title{The percentage each taxon makes up of a sample}
\usage{
seq_pct_taxa(Long, Taxon.List = Taxon_List)
}
\arguments{
\item{Long}{= Data in a long data format}

\item{Taxon.List}{= Master taxa list}
}
\value{
The percent of the sample represented by each taxa per taxon level.
}
\description{
The percentage each taxon makes up of a sample
}
