% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_metrics.R
\name{specific_diversity}
\alias{specific_diversity}
\title{Run All Diversity Metrics}
\usage{
specific_diversity(master.df, long.df, metrics.vec, class.wide, order.wide,
  family.wide, tribe.wide, genus.wide, taxa.rank.df, taxa.rank)
}
\arguments{
\item{master.df}{= The master taxa list containing all of the appropriate
taxonomic resolutions and the appropriate taxa attributes. A master taxa list
is contained with the BIBI package.}

\item{long.df}{= Taxonomic data in long format}

\item{metrics.vec}{= the vector of specified metrics to calculate.}

\item{class.wide}{= Class level wide taxonomic data frame}

\item{order.wide}{= Order level wide taxonomic data frame}

\item{family.wide}{= Family level wide taxonomic data frame}

\item{taxa.rank.df}{= Specified taxonomic level wide taxonomic data frame}

\item{taxa.rank}{= Taxonomic level ("FAMILY" or "GENUS") (Default = "FAMILY")}
}
\value{
All diversity metric values
}
\description{
Run All Diversity Metrics
}
