% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_metrics.R
\name{specific_non_diversity}
\alias{specific_non_diversity}
\title{Run All Non-Diversity Metrics}
\usage{
specific_non_diversity(master.df, long.df, metrics.vec, class.wide, order.wide,
  family.wide, tribe.wide, genus.wide, taxa.rank.df, taxa.rank)
}
\arguments{
\item{master.df}{= Taxonomic Information}

\item{long.df}{= Taxonomic data in long format}

\item{metrics.vec}{= the vector of specified metrics to calculate.}

\item{class.wide}{= Class level wide taxonomic data frame}

\item{order.wide}{= Order level wide taxonomic data frame}

\item{family.wide}{= Family level wide taxonomic data frame}

\item{taxa.rank.df}{= Specified taxonomic level wide taxonomic data frame}

\item{taxa.rank}{= Taxonomic level ("FAMILY" or "GENUS") (Default = "FAMILY")}
}
\value{
All non-diversity metric values
}
\description{
Run All Non-Diversity Metrics
}
