% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrimination_efficiency.R
\name{sse}
\alias{sse}
\title{Optimal Discrimination Efficiency}
\usage{
sse(metrics.df, quant.df, upper.class, lower.class, ref.df, quant.ref)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station with site
a column of site classes defined by environmental variables.}

\item{quant.df}{= Data frame containing upper.class quantile values.}

\item{upper.class}{= the site class that represents the better condition.}

\item{lower.class}{= the site class that represents the poorer condition.}

\item{ref.df}{= a data frame of only the upper.class values.}

\item{quant.ref}{= a data frame of reference quantile values.}
}
\value{
Determines the threshold at which a metric best categorizes
reference and degraded stations.
}
\description{
Optimal Discrimination Efficiency
}
