% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibi_methodologies.R
\name{test_ibi_method}
\alias{test_ibi_method}
\title{Run Multiple IBI Methodologies}
\usage{
test_ibi_method(raw.data, master.df, ibi.method, todays_date, redundant,
  range.var, zero.inflate, metric.groups, fam.group, jack.iterate,
  metric.class = m.c)
}
\description{
Run Multiple IBI Methodologies
}
