% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics _final.R
\name{tight_gradient}
\alias{tight_gradient}
\title{Tight Gradient Scoring}
\usage{
tight_gradient(metrics.df)
}
\arguments{
\item{metrics.df}{= a data frame containing raw metric values.}
}
\value{
Update
}
\description{
Tight Gradient Scoring
}
