% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_metrics _final.R
\name{uneven_gradient}
\alias{uneven_gradient}
\title{Uneven Gradient Scoring}
\usage{
uneven_gradient(metrics.df)
}
\arguments{
\item{metrics.df}{= a data frame containing raw metric scores.}

\item{metrics.summary}{= a data frame of the output from the function 
metric_summary.}

\item{bioregion}{= the bioregion of interest.}
}
\value{
Update
}
\description{
Uneven Gradient Scoring
}
