% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sensitivity.R, R/scoring_metrics2.R
\name{zero_inflate}
\alias{zero_inflate}
\alias{zero_inflate}
\title{Zero Inflation Test}
\usage{
zero_inflate(metrics.df, bi.barbour)

zero_inflate(metrics.df, bi.barbour)
}
\arguments{
\item{metrics.df}{= data frame of metric values for each station with site
a column of site classes defined by environmental variables.}

\item{bi.barbour}{= a data frame created within another function and used for
the calculated disturbance value.}

\item{metrics.df}{= data frame of metric values for each station.}

\item{bi.barbour}{= binary barbour output.}
}
\value{
Tests the influence of zeros on the results.

Scores the samples.
}
\description{
Zero Inflation Test

Zero Inflation
}
